Refactor SmtLibSolverBinary interface to allow multiple answers for commands (in case of interpolation)
